Move Block to components/attention_blocks.py
The dense transformer blocks (DropPath, Mlp, Attention_rel, Block_rel,
Block) are now complete in their own module.
